panion/requirements.txt
# Core Dependencies
aiohttp>=3.8.0
asyncio>=3.4.3
python-dateutil>=2.8.2
pyyaml>=6.0.0
redis>=4.5.0
packaging>=23.0
nest_asyncio>=1.5.8

# AI and Machine Learning
openai>=1.0.0
scikit-learn>=0.24.2
transformers>=4.0.0
nltk>=3.8.1
spacy>=3.7.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tensorflow>=2.6.0
torch>=1.9.0

# System and Infrastructure
pyzmq>=25.1.1
psutil>=5.9.0
netifaces>=0.11.0
pika>=1.3.2

# Data Processing and Storage
pandas>=2.1.3
sqlalchemy>=1.4.0
aiofiles>=0.8.0

# Async and Concurrency
apscheduler>=3.8.0
schedule>=1.1.0

# Web Framework and API
flask>=3.0.0
flask-swagger-ui>=4.11.0
apispec>=5.1.0
apispec-webframeworks>=0.5.0
marshmallow>=3.14.0

# Web Scraping and Automation
beautifulsoup4==4.12.3
selenium>=4.15.2
requests>=2.31.0
playwright>=1.40.0

# Security and Authentication
cryptography>=3.4.0
pyjwt>=2.0.0
pyotp>=2.6.0
qrcode>=7.3.0

# GUI and Media
PyQt6>=6.4.0
qtawesome>=1.2.0
pyautogui==0.9.54
PyYAML==6.0.1
opencv-python>=4.8.1.78
pillow>=10.1.0
pyperclip>=1.8.2

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.3.0
networkx>=2.6.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0
autopep8>=1.6.0
pylint>=2.12.0

# Utilities
jinja2>=3.1.2
pydantic>=2.5.2
jsonschema>=4.0.0
semver>=3.0.2
python-dotenv>=1.0.0
pluggy>=1.0.0
ttkbootstrap>=1.10.1

# Guidance
guidance>=0.1.0
pydantic>=2.0.0

# New dependencies
textblob==0.17.1
networkx==3.2.1
semver==3.0.2
python-dotenv==1.0.1
psutil==5.9.8
pydantic==2.6.3
injector==0.21.0 